+++ identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/repository/UserRepository.java
@@ -4,8 +4,8 @@ import nz.ac.canterbury.seng302.identityprovider.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface UserRepository  extends CrudRepository <User, Long> {
+public interface UserRepository  extends CrudRepository <User, Integer> {
     User findByUsername(String username);
 
-    User findByUserId(Long userId);
+    User findByUserId(Integer userId);
 }
